+++ src/data/skill/public/init.lua
@@ -1,0 +1,4 @@
+return require 'data.load_file'("data.skill.public.", {
+    "test",
+    "test1"
+})