+++ src/skill_tree/public/init.lua
@@ -1,4 +1,0 @@
-local require = require
-
-require 'skill_tree.public.test'
-require 'skill_tree.public.test1'